backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
mysql-connector-python==8.3.0
reportlab==4.2.2
num2words==0.2.13
pdfkit==1.0.0
qrcode==7.4.2
Pillow==10.0.0